requirement.txt
pytorch
torchvision
torchnet
tqdm
fire
